requirements/base.txt
Django==1.8.1
Pillow==2.8.1

django-bootstrap3==5.4.0
django-extensions==1.5.3
django-filter==0.10.0
django-redis==4.0.0
djangorestframework==3.1.1
psycopg2==2.6
python-social-auth==0.2.7
sorl-thumbnail==12.2
gunicorn==19.3
raven==5.3.1